% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2var.R
\name{check_2var}
\alias{check_2var}
\title{Check plot_2var}
\usage{
check_2var(data, x, y)
}
\arguments{
\item{data}{See plot_2var}

\item{x}{See plot_2var}

\item{y}{See plot_2var}
}
\description{
Check plot_2var
}
