% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{missing_method}
\alias{missing_method}
\title{You better call Saul}
\usage{
missing_method(f, ...)
}
\arguments{
\item{f}{function with missing method}

\item{...}{missing class method}
}
\description{
You better call Saul
}
