% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2var.R
\name{plot_2var}
\alias{plot_2var}
\title{Bivariate Plots}
\usage{
plot_2var(data, x = NULL, y = NULL, remove = NULL, ...)
}
\arguments{
\item{data}{A data frame}

\item{x}{Strings. Explicatory variables. If NULL: names(data)}

\item{y}{Strings. Target variable. If NULL: last x.}

\item{remove}{Strings. Variables to exclude.}

\item{...}{Extra parameters for methods}
}
\value{
Returns invisibly a bivariate plot or a named list of bivariate plots, ggplot2 objects
}
\description{
Bivariate plots accordig to variables type
}
\examples{

test <- data.frame(num_var  = rnorm(100),
                   num_var2 = rnorm(100),
                   chr_var  = sample(letters, 100, TRUE),
                   fct_var  = factor(sample(LETTERS, 100, TRUE)),
                   dat_var  = seq.Date(Sys.Date(), by = "day", length.out = 100),
                   pos_var  = as.POSIXct(as.character(seq.Date(Sys.Date(), by = "day", length.out = 100))),
                   stringsAsFactors = FALSE)

plot_2var(test, y = "num_var")
plot_2var(test, y = "chr_var")
plot_2var(iris)
# automatically sets x to names of data and y to species

}
