% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_bar}
\alias{plot_bar}
\title{Barplot}
\usage{
plot_bar(data, x, y, ...)
}
\arguments{
\item{data}{a data frame or tibble}

\item{x}{string. factor variable to colour the graph}

\item{y}{string. factor variable to show the count}
}
\value{
a barplot, ggplot2 object
}
\description{
Plots a standard barplot using ggplot2
}
\examples{

df <- mtcars
df$cyl <- factor(df$cyl)
df$gear <- factor(df$gear)

plot_bar(df, "cyl", "gear")

}
