% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Boxplot}
\usage{
plot_boxplot(data, x, y, ...)
}
\arguments{
\item{data}{a data frame or tibble}

\item{x}{string. numeric or integer variable}

\item{y}{string. factor variable}
}
\value{
a boxplot, ggplot2 object
}
\description{
Plots a standard boxplot using ggplot2
}
\examples{

plot_boxplot(iris, "Petal.Length", "Species")

}
