% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Scatter plot}
\usage{
plot_scatter(data, x, y, ...)
}
\arguments{
\item{data}{a data frame or tibble}

\item{x}{string. a numeric variable}

\item{y}{string. a numeric variable}
}
\value{
a barplot, ggplot2 object
}
\description{
Plots a scatter plot between two variables using ggplot2
}
\examples{

plot_scatter(iris, "Petal.Length", "Petal.Width")

}
