% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_var.R
\name{plot_var}
\alias{plot_var}
\title{Histogram or Bar Chart for each variable}
\usage{
plot_var(x, ...)
}
\arguments{
\item{x}{Vector, List or Dataframe.}

\item{...}{Additional parameters (see details)}
}
\value{
Returns a plot or a list of plots, ggplot2 objects. Automatically plots all of them.
}
\description{
Plot a vector, each variable of a list, or each variable of a dataframe
}
\details{
lab: name to write
}
\examples{

# vector
plot_var(iris$Species, "Species")
plot_var(iris$Petal.Length, "Petal.Length")

# dataframe
plot_var(iris)

# tibble
plot_var(tibble::as_tibble(iris))

# list
plot_var(list(rnorm(1000),
       rcauchy(1000),
       rpois(1000, 100)))

# named list
plot_var(list(norm = rnorm(1000),
       cauchy = rcauchy(1000),
       poisson = rpois(1000, 100)))

# list of lists and dataframes
plot_var(list(iris,
       list(rnorm(1000),
            rcauchy(1000),
            rpois(1000, 100))))

}
