% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2var.R
\name{plot_2var}
\alias{plot_2var}
\title{Bivariate Plots}
\usage{
plot_2var(data, x, y, ..., remove = NULL)
}
\arguments{
\item{data}{a data frame or tibble}

\item{x}{an explicative variable}

\item{y}{target variable}

\item{...}{extra parameters for methods}

\item{remove}{a character vector of columns to exclude from plotting}
}
\value{
Returns invisibly a bivariate plot or a named list of bivariate plots, ggplot2 objects
}
\description{
Bivariate plots accordig to variables type
}
\examples{

test <- data.frame(num_var = rnorm(100),
                   chr_var = sample(letters, 100, TRUE),
                   fct_var = factor(sample(LETTERS, 100, TRUE)),
                   dat_var = seq.Date(Sys.Date(), by = "day", length.out = 100),
                   pos_var = as.POSIXct(as.character(seq.Date(Sys.Date(), by = "day", length.out = 100))),
                   stringsAsFactors = FALSE)



lapply(combn(names(test), 2, simplify = FALSE),
       function(x, test) plot_2var(test, x[1], x[2]), test)
lapply(combn(names(test), 2, simplify = FALSE),
       function(x, test) plot_2var(test, x[2], x[1]), test)


plot_2var(iris, setdiff(names(iris), "Species"), "Species")
plot_2var(iris, setdiff(names(iris), "Species"), "Species",
 remove = names(iris)[tidyselect::starts_with("Petal", vars = names(iris))])

}
